0.2
Added the IDgenerator class to create random 16 char long strings to uniquely identify each node.

Added some comments and documentation.

Wrote most of the code for the Node class (except the run function)

Changed the arguments of gates from arrays to vectors.

Deleted the static class Gates, the functions it contained are now on their own.