0.5
Added an array containing the inputs and outputs ID's (useful when building a network).

Changed the structure of the nodes in a network: It is now a 2 dimensional array. The number of inputs, outputs and layers can be derived from it so these class parameters have been deleted and replaced by getter functions.

Rewrote the neuralnetworc constructor to fit the new architecture.

Started making a second constructor to make new networks instead of loading previous ones