0.7
Added 3 NeuralNetwork.run() methods, allowing the networks to be finally useables.

The first one, run() without arguments, simply updates the value of every node from first to last layer (except input layer) and returns the values of th elast layer (The Network's outputs)

The second one, run(int index, bool value), allows the user to update a single input and run only the Nodes that have been changed to save ressources when using a network live.

The last one, run(std::vector<bool> inputs) takes an array of as many elements as there are inputs in the network, sets the inputs to be that and runs it, returning the resulting output

Added the NeuralNetwork.getOutputs() and NeuralNetwork.setInputs() getters and setters

Updated the Node.evaluate() method to work when there are no inputs to the node (simply returns the current output)